Remove the ResumeSessionHandler class
