+++ Main.cc
@@ -26,7 +26,6 @@
 #include "FileListHandler.h"
 #include "FileSettings.h"
 #include "OnMessageTask.h"
-#include "ResumeSessionHandler.h"
 #include "Session.h"
 #include "Util.h"
 
@@ -155,7 +154,7 @@ void OnMessage(uWS::WebSocket<uWS::SERVER>* ws, char* raw_message, size_t length
                 case CARTA::EventType::RESUME_SESSION: {
                     CARTA::ResumeSession message;
                     if (message.ParseFromArray(event_buf, event_length)) {
-                        ResumeSessionHandler(session, message, head.request_id);
+                        session->OnResumeSession(message, head.request_id);
                     } else {
                         fmt::print("Bad RESUME_SESSION message!\n");
                     }
